RAGUsingOpenAIGPT4.0/requirement.txt
# openai==1.37.2
openai==1.72.0
python-dotenv==1.0.1